Mvaldes/feature/login (#16)
* ➕ added a authForm to signIn page

Followed the tutorial
https://supertokens.com/blog/building-a-login-screen-with-react-and-bootstrap
but mostly the doc https://react-bootstrap.github.io/forms/overview/ to
create a sign in, sign up form in bootstrap react.

* 🎨 WIP work on styling bootstrap components

WIP to find out how to customize bootstrap components.

* 🎨 sign up form style change for form only

Found a way to customize bootstrap and change colors shapes etc for the
form. Will be tacling buttons right after.

* 🚧 wip on button style

Only changed the color, trying to change the size

* 🚧 wip button styling

Button are almost styled, issues with buttons border and states of
buttons.

* 🐛 stretched link issue

Used a stretched link for sign in / sign up that was not useful.

* 🚧 wip to add a onSubmit action

Wip for try and find a way to get the elements from the form and send
them to the back.

* ➕ sign in / sign up front handleSubmit

Can fetch data from the form. Next step is to send data to the back.

* 🚧 wip post request to back

Trying to post using fetch to http://localhost:4000/auth/signup when
using the submit button.

* 🚧 WIP change routes for auth + enabled cors

enabled Cors in back folder + Wip changing routes

* ➕ auth methods as childs of auth in router

➕ routes signup and signin are now childs of route auth: when we sitch between auth/signin and auth/signup, only the differences will render.
➕ Signin and signup form now have their own routes.
➕ Going to /auth will redirect to /auth/signin

* 🧹 cleanup of game files and deleted route example

🗑️ deleted the route example for leaderboard
🧹 put the game files in the route folder as it is a route

* ➕ userinfo added to localStor. wip log in context

Userinfo are stored in local storage with usertoken.
Wip setting up context for login status

* ➕ signIn and signup fct for fetch

Distinguished signn and signup function to fetch and store token + data.
Will work on redirecting to protected page after.

* 🚧 wip protect a page with login

Following :
https://stackblitz.com/github/remix-run/react-router/tree/main/examples/auth?file=src%2FApp.tsx,src%2Fmain.tsx,src%2Fauth.ts
to try to learn how to protect a page with login status.

* ➕ login to access protected pages

You now have to login to access protected pages.
The Auth function, on submit:
- call to the API
- if token is received, store userInfo to localStorage
- set context AuthContext auth.user with userEmail (for now, has to be
changed later to username)
- display on top of the page the userEmail and a button to signOut

Signout :
- clears the localStorage
- reset auth.user to null
- navigate to "/", the home page

if user tries to access protected page while not signed in:
- he'll be redirected to sign in page
- he'll be then redirected to protected page

* 🐛 fix login after signup and sign out then sign in

fixed a bug when user sign up, then sign out, then tries to sign in
again.

* 🐛 changed email to username for SignIn

*