Mvaldes/feature/user private profile (#25)
* 🚧 wip routes for protected pages of the app

*

* ➕ flow from route/route from landing to frst login

Changed the flow to be redirected from landing page, to first login

* 🚧 building userprofile cards

*

* ➕ nested restricted page under app/

➕ all restricted page are now under app/ & require to login
🗑️ game page that is not needed

* 🚧 user private profile cards wip

working on the css

* 🚧 wip css for userprofile

*

* ➕ routes changes for SPA style

Plus cleanup of some files

* :construction: wip front

*

* 🚧 progress on CSS

*

* ➕ first user profile card done

Displays user information with buttons to edit those info.
Next step is to open a second card when edit button is clicked to modify
said info.

* ➕ card that show up to modify entries

➕ a card that pops when we want to edit user info
🔨 prettier to other files

* ➕ modify card use hooks onChange

The right card used to modify the user info, updates the user entry
onChange, and will be able to send the data to the back.

next step is to validate the input before making fetch request to the
back!

* 🚧 updating in back and front new entries

Trying to fetch a post to the back to update users info on submit.
Updating front display of userInfo on the front.

Currently both work, just not together.

* ➕ update username in front and back

But not the other entries yet.

* ➕ updates email front and back

plus, separated concerns with queries

* 🔨 files structure refacto

folder for hooks, globals, and queries to clean up code and separate
concerns

* ➕ webpack config for sgv and jpeg

used to load imgs and icons to html

* 🐛 signup was calling signin

*

* ➕ u profile icons and images

added icons and images to user private profile.
the user avatar will need to be loaded to DB and then displayed. For
now, it is a placeholder

* ➕ added modal on click to update avatar

Added a modal that pops up when click the edit button to add or update
the user avatar.

* ➕ POST avatar query

*

* 🚧 users list

*

* 🚧 list of friends and blocked users

* ➕ user still logged in on refresh

added the user context to local storage so that, on refresh, the context
persists.

* 🧹 organized auth files

placed auth hooks and context in their corresponding files and folders.

* 🚧 promises for auth

*